Conversion of int to Long

+++ ERP_V2/src/main/java/com/example/ERP_V2/DTO/OrderDTO.java
@@ -21,7 +21,7 @@ public class OrderDTO {
 
     private int orderId;
     private Date date;
-    private BigDecimal estimatedAmount;
+    private Long estimatedAmount;
     private String paperSize;
     private int pages;
     private int quantity;
